+++ patches/src/main/kotlin/io/github/chmate/revanced/ChMatePatch.kt
@@ -14,6 +14,7 @@ val chMatePatch = bytecodePatch(
     extendWith("extensions/chmate.rve")
 
     apply {
+        patchIntegrityChecks()
         patchNetworkBoundaries()
         renameSanitizedManifestClasses()
     }